requirements.txt
appnope==0.1.0
bleach==2.1.1
cycler==0.10.0
Cython==0.27.3
decorator==4.1.2
dill==0.2.7.1
docutils==0.14
entrypoints==0.2.3
html5lib==1.0b10
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.3
jedi==0.11.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.7.4
multidict==3.3.2
multiprocess==0.70.5
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.0
numpy==1.13.3
pandas==0.20.3
pandocfilters==1.4.2
parso==0.1.0
pathos==0.2.1
pexpect==4.2.1
pickleshare==0.7.4
pox==0.2.3
ppft==1.6.4.7.1
progressbar2==3.34.3
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-utils==2.2.0
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.1
readme-renderer==17.2
restview==2.7.0
scipy==1.0.0
seaborn==0.8.1
simplegeneric==0.8.1
six==1.11.0
terminado==0.6
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.0.6
